+++ polaris-react/src/components/LegacyCard/utilities/index.ts
@@ -1,0 +1,1 @@
+export {useLegacyCardPaddingObserverRef} from './use-legacy-card-padding-observer-ref';
